allow "raw" data to be passed in to a cURL request
this adds a new encoding type, ENCODING_RAW, to the Request class. This
encoding type takes the POST data "as is", merely ensuring it is a
string prior to passing it along to cURL itself.

this commit removes the requirement that setData() be given an array. A
raw string, or any stringable data, may be passed as well.

this commit adds magic raw*() methods to the cURL class, to complement
the json*() magic methods, as well as a "newRawRequest()" helper method.